Fixes a couple compiler warnings.
git-svn-id: http://svn.mythtv.org/svn/trunk@22004 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/DeviceReadBuffer.cpp
@@ -374,7 +374,7 @@ bool DeviceReadBuffer::Poll(void) const
         }
         else //  ret == 0
         {
-            if (timer.elapsed() > max_poll_wait)
+            if ((uint)timer.elapsed() > max_poll_wait)
             {
                 VERBOSE(VB_IMPORTANT, LOC_ERR + "Poll giving up");
                 QMutexLocker locker(&lock);
